requirements.txt
bs4==0.0.1
python-vlc==3.0.11115
requests==2.25.0
m3u8==0.7.1


